+++ code/c/common/opdi_message.c
@@ -153,9 +153,9 @@ static uint8_t decode(opdi_Message *message, uint8_t bytes[]) {
 */
 static uint8_t encode(opdi_Message *message, uint16_t *length) {
 	char channelBuf[CHANNEL_MAXBUF + 1] = {'\0'};
-	uint32_t pos = 0;
+	size_t pos = 0;
 	uint16_t checksum = 0;
-	uint16_t i;
+	size_t i;
 	uint8_t err;
 	uint16_t bytelen = 0;
 	uint8_t nibble;
